Add public headers to KSDeferred.h, turn on werror for the ios specs
Basically, if you use this as a dynamic framework (e.g. carthage), and have werror enabled
Then the compiler will error (generate a warning, actually) because it can't find headers
which should be publicly available in the main public header (e.g. KSDeferred.h).

This is also not something the tests will catch on, likely because the framework is successfully built
and then the test bundle is injected into the framework.

Maybe if we were using a test suite for this, then it might catch that?