Don’t return `nil` from `then`
The header of `KSPromise` defines `then:` and `then:error:`
to return a `nonnull` value but the implementation can return
`nil` when calling one of these methods on a cancelled promise.
This especially is problematic for interoperability with Swift since
it relies on the nullability annotations.

This changes the behavior so that chaining a promise to a
cancelled promise instantly cancels the returned promise
turning it basically into a noop.